requirements.txt
web3
flask
python-dotenv 